+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
   * (theimo1221) Add option to run specific ac's regardless of energy level
+  * (theimo1221) Add option to disable dachs altogether
 -->
 ## **WORK IN PROGRESS**
 * (theimo1221) Fix manual shutter down in evening hours messing up automatic movement